splitting CONTACT.js into 2 components
